requirements.txt
streamlit
python-dotenv
pdf2image
pytesseract
pymupdf
groq
google-api-python-client
